requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
cycler==0.11.0
Django==4.0.1
djangorestframework==3.13.1
fonttools==4.28.5
gunicorn==20.1.0
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.1
packaging==21.3
pandas==1.3.5
Pillow==9.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.0.0
whitenoise==5.3.0
